requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
click==8.1.3
fastapi==0.86.0
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-decouple==3.6
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
